feat: use local product images in cart LineItem (resolves #18) (#128)
This adds a `ProductImage` component that takes an image ID and returns a Gatsby Image dressed and ready to head to the dance (i.e., mapped back to a local, responsive image). This resolves/addresses #18.

Cool note: this typically results in the browser selecting the same image that was used for the product preview since it's already loaded, resulting in zero additional requests/data. Win!