chore: update react-icons to latest v3 version. (#145)
# Overview
Closes #144 (typo in branch name). 
Update react-icons to v3. Small fixes for v2 to v3 upgrade. 

# Questions
Are we happy with placing the `IconContext.Provider` in the Header component? I was thinking it could go in layout, but wasn't sure if that was too heavy-handed. 
Let me know if you want to move and I can make the update.