refactor: move auth from local storage to memory (#265)
This is a PR to move the Auth0 tokens from local storage into memory. It also wraps the root element in a `SessionCheck` component to handle silent authentication if the user is already logged in.

This PR Is in DRAFT status because I am unable to debug some of the errors I'm receiving locally. This is likely due to settings in the Auth0 application, e.g. `Allowed Web Origins`, that I don't have access to. If a Gatsby team member could review this code and try to debug it/update the Auth0 application settings, I think we're nearly there with this code!